Fix leftover from #8236 and make BufferReader/Writer support Span
- Fix leftover nits from @bzbarsky-apple's review of #8236
- In order to add span support cleanly, added Span support to
  Reader and BufferWriter, and fixed all necessary breakage.

Testing done: pass all unit tests and CASE cert tests

+++ src/protocols/bdx/BdxTransferSession.h
@@ -52,7 +52,7 @@ public:
 
         // Additional metadata (optional, TLV format)
         const uint8_t * Metadata = nullptr;
-        uint16_t MetadataLength  = 0;
+        size_t MetadataLength    = 0;
     };
 
     struct TransferAcceptData
@@ -65,7 +65,7 @@ public:
 
         // Additional metadata (optional, TLV format)
         const uint8_t * Metadata = nullptr;
-        uint16_t MetadataLength  = 0;
+        size_t MetadataLength    = 0;
     };
 
     struct StatusReportData
@@ -76,7 +76,7 @@ public:
     struct BlockData
     {
         const uint8_t * Data = nullptr;
-        uint16_t Length      = 0;
+        size_t Length        = 0;
         bool IsEof           = false;
     };
 
@@ -314,7 +314,7 @@ private:
     TransferAcceptData mTransferAcceptData;
     BlockData mBlockEventData;
 
-    uint32_t mNumBytesProcessed = 0;
+    size_t mNumBytesProcessed = 0;
 
     uint32_t mLastBlockNum = 0;
     uint32_t mNextBlockNum = 0;
